refactor: examples and tests
