--- examples/autoscale/README.md
@@ -1,8 +1,0 @@
-# Autoscale example
-
-An end-to-end example that uses the module's default variable values.
-
-This example uses the IBM Cloud Terraform provider to create the following infrastructure:
-
-- A resource group, if one is not passed in.
-- An ICD MongoDB database instance with autoscaling enabled (automatically increase resources).
