+++ examples/basic/main.tf
