+++ examples/basic/variables.tf
