+++ examples/complete/README.md
@@ -1,9 +1,9 @@
 # Complete example with Encryption and CBR rules
 
-An end-to-end example that adds encryption to the [default example](../default/README.md). This example uses the IBM Cloud terraform provider to:
+An end-to-end example that adds encryption to the [default example](../basic/README.md). This example uses the IBM Cloud terraform provider to:
 
 - Create a new resource group if one is not passed in.
-- Create a new mongoDB database instance.
+- An ICD MongoDB database instance with autoscaling enabled (automatically increase resources).
 - Create Key Protect instance with root key.
 - Backend encryption using generated Key Protect key.
 - Create a Sample VPC.
