+++ tests/pr_test.go
@@ -17,8 +17,6 @@ const regionSelectionPath = "../common-dev-assets/common-go-assets/icd-region-pr
 // Allow the tests to create a unique resource group for every test to ensure tests do not clash. This is due to the fact that the auth policy created by this module has to be scoped to the resource group and hence would clash if tests used same resource group.
 //const resourceGroup = "geretain-test-mongo"
 
-const defaultExampleTerraformDir = "examples/default"
-const autoscalingExampleTerraformDir = "examples/autoscale"
 const completeExampleTerraformDir = "examples/complete"
 const fsCloudTerraformDir = "examples/fscloud"
 
@@ -42,35 +40,6 @@ func TestMain(m *testing.M) {
 	os.Exit(m.Run())
 }
 
-func TestRunDefaultExample(t *testing.T) {
-	t.Parallel()
-
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:      t,
-		TerraformDir: defaultExampleTerraformDir,
-		Prefix:       "mongodb-def",
-	})
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-func TestRunAutoscaleExample(t *testing.T) {
-	t.Parallel()
-
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:            t,
-		TerraformDir:       autoscalingExampleTerraformDir,
-		Prefix:             "mongodb-auto",
-		BestRegionYAMLPath: regionSelectionPath,
-	})
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
 func TestRunFSCloudExample(t *testing.T) {
 	t.Parallel()
 	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
@@ -88,33 +57,7 @@ func TestRunFSCloudExample(t *testing.T) {
 	assert.NotNil(t, output, "Expected some output")
 }
 
-func testRunComplete(t *testing.T, version string) {
-	t.Parallel()
-
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:            t,
-		TerraformDir:       completeExampleTerraformDir,
-		Prefix:             "mongodb-comp",
-		BestRegionYAMLPath: regionSelectionPath,
-		TerraformVars: map[string]interface{}{
-			"mongodb_version": version,
-		},
-	})
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-func TestRunCompleteExample(t *testing.T) {
-	t.Parallel()
-	versions := []string{"4.2", "4.4"}
-	for _, version := range versions {
-		t.Run(version, func(t *testing.T) { testRunComplete(t, version) })
-	}
-}
-
-func TestRunUpgradeExample(t *testing.T) {
+func TestRunCompleteUpgradeExample(t *testing.T) {
 	t.Parallel()
 
 	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
@@ -122,6 +65,9 @@ func TestRunUpgradeExample(t *testing.T) {
 		TerraformDir:       completeExampleTerraformDir,
 		Prefix:             "mongodb-upg",
 		BestRegionYAMLPath: regionSelectionPath,
+		TerraformVars: map[string]interface{}{
+			"mongodb_version": "4.2",
+		},
 	})
 
 	output, err := options.RunTestUpgrade()
